Network.URI.Encode

Description

Unicode aware URI encoding and decoding functions for both String and Text. Although standards are pretty vague about Unicode in URIs most browsers are pretty straightforward when encoding URIs: Encode a text to a UTF-8 string and URI-encode every individual byte (not character).

Documentation

encode :: String -> String

URI encode a String, unicode aware.

encodeWith :: (Char -> Bool) -> String -> String

URI encode a String, unicode aware, using the predicate to decide which characters are escaped (False means escape).

encodeText :: Text -> Text

URI encode a Text, unicode aware.

encodeTextWith :: (Char -> Bool) -> Text -> Text

URI encode a Text, unicode aware, using the predicate to decide which characters are escaped (False means escape).

encodeTextToBS :: Text -> ByteString

URI encode a Text into a ByteString, unicode aware.

encodeTextToBSWith :: (Char -> Bool) -> Text -> ByteString

URI encode a Text into a ByteString, unicode aware, using the predicate to decide which characters are escaped (False means escape).

encodeByteString :: ByteString -> ByteString

URI encode a UTF8-encoded ByteString into a ByteString, unicode aware.

encodeByteStringWith :: (Char -> Bool) -> ByteString -> ByteString

URI encode a UTF8-encoded 'ByteString into a ByteString, unicode aware, using the predicate to decide which characters are escaped (False means escape).

decode :: String -> String

URI decode a String, unicode aware.

decodeText :: Text -> Text

URI decode a Text, unicode aware.

decodeBSToText :: ByteString -> Text

URI decode a ByteString into a Text, unicode aware.

decodeByteString :: ByteString -> ByteString

URI decode a ByteString into a UTF8-encoded ByteString, unicode aware.

isAllowed :: Char -> Bool

Is a character allowed in a URI. Only ASCII alphabetic characters, decimal digits, and - _ . ~ are allowed. This is following RFC 3986.